Move collection into its own directory
