+++ frog/imports/client/GraphEditor/SidePanel/OperatorPanel/ChooseOperator.jsx
@@ -15,7 +15,7 @@ import TextField from '@material-ui/core/TextField';
 import Tooltip from '@material-ui/core/Tooltip';
 import Collapse from '@material-ui/core/Collapse';
 import { type operatorPackageT, type OperatorDbT, cloneDeep } from 'frog-utils';
-import { Operators } from '/imports/api/operators';
+import { Operators } from '/imports/collections';
 import { operatorTypes, operatorTypesObj } from '/imports/operatorTypes';
 import { type StoreProp } from '../../store';
