+++ frog/imports/client/GraphEditor/TopPanel/GraphConfigPanel.jsx
@@ -5,7 +5,8 @@ import { withTracker } from 'meteor/react-meteor-data';
 import { ChangeableText } from 'frog-utils';
 
 import { connect, store } from '../store';
-import { Graphs, renameGraph } from '/imports/api/graphs';
+import { Graphs } from '/imports/collections';
+import { renameGraph } from '/imports/api/graphs';
 import { ValidButton } from '../Validator';
 
 const Config = ({ graph }) => (
