+++ frog/imports/client/StudentView/SessionBody.jsx
@@ -16,9 +16,8 @@ import { getInitialState, withDragDropContext } from 'frog-utils';
 import { compose, toClass } from 'recompose';
 import { withRouter } from 'react-router';
 
-import { Activities } from '/imports/api/activities';
+import { Activities, Sessions } from '/imports/collections';
 import { logLogin } from '/imports/api/logs';
-import { Sessions } from '/imports/api/sessions';
 import Runner from './Runner';
 import Countdown from './Countdown';
