+++ frog/imports/client/TeacherView/Roster.jsx
@@ -20,7 +20,7 @@ import PersonIcon from '@material-ui/icons/Person';
 import AddIcon from '@material-ui/icons/Add';
 import Delete from '@material-ui/icons/Delete';
 import { withStyles } from '@material-ui/core/styles';
-import { Sessions } from '/imports/api/sessions';
+import { Sessions } from '/imports/collections';
 
 import styles from './styles';
