+++ frog/imports/client/TeacherView/utils/exportComponent.js
@@ -7,10 +7,13 @@ import FileSaver from 'file-saver';
 import { cleanEmptyCols, type ActivityDbT, strfTime } from 'frog-utils';
 import { omit } from 'lodash';
 
-import { UploadList, downloadFile } from '/imports/api/openUploads';
-import { Sessions } from '/imports/api/sessions';
-import { Activities } from '/imports/api/activities';
-import { Operators } from '/imports/api/operators';
+import {
+  UploadList,
+  Sessions,
+  Activities,
+  Operators
+} from '/imports/collections';
+import { downloadFile } from '/imports/api/openUploads';
 import { activityTypesObj } from '/imports/activityTypes';
 import { getActivitySequence } from '/imports/api/graphSequence';
 import { graphToString } from '/imports/api/exportGraph';
