+++ frog/server/externalOperators.js
@@ -2,7 +2,7 @@
 import redis from 'redis';
 import { Meteor } from 'meteor/meteor';
 import { operatorTypes, operatorTypesObj } from '../imports/operatorTypes';
-import { ExternalOperators } from '../imports/api/operators';
+import { ExternalOperators } from '/imports/collections';
 
 export default () => {
   ExternalOperators.find({})
