+++ frog/server/getLogMethods.js
@@ -1,8 +1,7 @@
 // @flow
 import { Meteor } from 'meteor/meteor';
 
-import { Logs } from '../imports/api/logs';
-import { Sessions } from '../imports/api/sessions';
+import { Logs, Sessions } from '/imports/collections';
 import { mergeLog } from '../imports/api/mergeLogData';
 import { client } from './redis';
 import { archiveDashboardState } from './dashboardSubscription';
