+++ frog/server/runDataflow.js
@@ -9,15 +9,18 @@ import {
 } from 'frog-utils';
 
 import { generateReactiveFn } from '/imports/api/generateReactiveFn';
-import { Sessions } from '/imports/api/sessions';
+import {
+  Sessions,
+  Products,
+  Activities,
+  Connections,
+  Operators
+} from '/imports/collections';
 import mergeData from './mergeData';
 import { serverConnection } from './share-db-manager';
 import reactiveToProduct from './reactiveToProduct';
 import { operatorTypesObj } from '../imports/operatorTypes';
 import operators from './operatorRunners';
-import { Products } from '../imports/api/products';
-import { Activities, Connections } from '../imports/api/activities';
-import { Operators } from '../imports/api/operators';
 import { addObject } from '../imports/api/objects';
 import remote from './runRemoteOperator';
