+++ frog/server/teacherImports.js
@@ -4,14 +4,15 @@ import { Meteor } from 'meteor/meteor';
 import {
   Activities,
   Connections,
-  DashboardData
-} from '../imports/api/activities.js';
-import { UploadList } from '../imports/api/openUploads.js';
-import { Operators, ExternalOperators } from '../imports/api/operators.js';
-import { Graphs } from '../imports/api/graphs.js';
-import { Sessions } from '../imports/api/sessions.js';
-import { Products } from '../imports/api/products.js';
-import { Objects } from '../imports/api/objects.js';
+  DashboardData,
+  UploadList,
+  Operators,
+  ExternalOperators,
+  Graphs,
+  Sessions,
+  Products,
+  Objects
+} from '/imports/collections';
 
 const teacherPublish = (publish, collection, limitation) => {
   Meteor.publish(publish, function() {
